:recycle: move definitions in lib
