+++ src/lib.rs
@@ -1,0 +1,2 @@
+mod ink;
+pub use ink::*;
